allow a terminal to be inited without a shell

+++ src/window.c
@@ -239,16 +239,19 @@ int DLLEXPORT ltm_set_window_dimensions(int tid, ushort width, ushort height) {
 	if(descs[tid].pid) {
 		if(tcsetwinsz(tid) == -1) return -1;
 
-		pgrp = tcgetpgrp(fileno(descs[tid].pty.master));
-		if(pgrp == -1) SYS_ERR("tcgetpgrp", NULL);
+		/* only do this stuff if there is actually a shell to signal */
+		if(descs[tid].pid > 0) {
+			pgrp = tcgetpgrp(fileno(descs[tid].pty.master));
+			if(pgrp == -1) SYS_ERR("tcgetpgrp", NULL);
 
-		/* not caring whether this fails or not
-		 * it can do so if there's currently no
-		 * foreground process group or if we don't
-		 * have permission to signal some processes
-		 * in the process group
-		 */
-		killpg(pgrp, SIGWINCH);
+			/* not caring whether this fails or not
+			 * it can do so if there's currently no
+			 * foreground process group or if we don't
+			 * have permission to signal some processes
+			 * in the process group
+			 */
+			killpg(pgrp, SIGWINCH);
+		}
 
 		/* don't do this before ltm_term_init since the callbacks
 		 * aren't guaranteed to be set (and there isn't much point
